Silence warnings in examples

+++ examples/readTest/readTest.ino
@@ -18,13 +18,13 @@ uint16_t OSFS::startOfEEPROM = 1;
 uint16_t OSFS::endOfEEPROM = 1024;
 
 void OSFS::readNBytes(uint16_t address, unsigned int num, byte* output) {
-	for (int i = address; i < address + num; i++) {
+	for (uint16_t i = address; i < address + num; i++) {
 		*output = EEPROM.read(i);
 		output++;
 	}
 }
 void OSFS::writeNBytes(uint16_t address, unsigned int num, const byte* input) {
-	for (int i = address; i < address + num; i++) {
+	for (uint16_t i = address; i < address + num; i++) {
 		EEPROM.update(i, *input);
 		input++;
 	}
